+++ CHANGES.md
@@ -30,6 +30,17 @@ To be released.
 [#3769]: https://github.com/planetarium/libplanet/pull/3769
 
 
+Version 4.4.2
+-------------
+
+Released on April 29, 2024.
+
+ -  (Libplanet.Explorer) Added `KeyBytesType` and `TrieType`.  [[#3763]]
+ -  (Libplanet.Explorer) Added `RawStateQuery`.  [[#3763]]
+
+[#3763]: https://github.com/planetarium/libplanet/pull/3763
+
+
 Version 4.4.1
 -------------